# Perf/value tuples