--- a/src/Bandersnatch/Nethermind.Field.Arithmetic.Benchmark/Benchmarks.cs
+++ b/src/Bandersnatch/Nethermind.Field.Arithmetic.Benchmark/Benchmarks.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using BenchmarkDotNet;
+using BenchmarkDotNet.Attributes;
+
+namespace Nethermind.Field.Arithmetic.Benchmark
+{
+    public class Benchmarks
+    {
+        [Benchmark]
+        public void Scenario1()
+        {
+            // Implement your benchmark here
+        }
+
+        [Benchmark]
+        public void Scenario2()
+        {
+            // Implement your benchmark here
+        }
+    }
+}