--- a/src/Bandersnatch/Nethermind.Field.Arithmetic.Benchmark/Nethermind.Field.Arithmetic.Benchmark.csproj
+++ b/src/Bandersnatch/Nethermind.Field.Arithmetic.Benchmark/Nethermind.Field.Arithmetic.Benchmark.csproj
@@ -14,7 +14,9 @@
         <IsPackable>false</IsPackable>
     </PropertyGroup>
     <ItemGroup>
-        <PackageReference Include="BenchmarkDotNet" Version="0.12.1" />
-        <PackageReference Include="BenchmarkDotNet.Diagnostics.Windows" Version="0.12.1" Condition="'$(OS)' == 'Windows_NT'" />
+        <PackageReference Include="BenchmarkDotNet" Version="0.13.4" />
+    </ItemGroup>
+    <ItemGroup>
+      <ProjectReference Include="..\Nethermind.Field.Montgomery\Nethermind.Field.Montgomery.csproj" />
     </ItemGroup>
 </Project>