--- a/src/Bandersnatch/Nethermind.Field.Montgomery.Test/ElementTests.cs
+++ b/src/Bandersnatch/Nethermind.Field.Montgomery.Test/ElementTests.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
-using Nethermind.Field.Montgomery.ElementFactory;
 using NUnit.Framework;
+using FE = Nethermind.Field.Montgomery.ElementFactory.Element;
 
 namespace Nethermind.Field.Montgomery.Test
 {
@@ -12,12 +12,12 @@ namespace Nethermind.Field.Montgomery.Test
         [Test]
         public void TestNegativeValues()
         {
-            using IEnumerator<Element> set = Element.GetRandom().GetEnumerator();
+            using IEnumerator<FE> set = FE.GetRandom().GetEnumerator();
             for (int i = 0; i < 1000; i++)
             {
-                Element x = set.Current;
-                Element y = 0 - x;
-                Element z = y + x;
+                FE x = set.Current;
+                FE y = 0 - x;
+                FE z = y + x;
                 Assert.IsTrue(z.IsZero);
                 set.MoveNext();
             }
@@ -26,36 +26,50 @@ namespace Nethermind.Field.Montgomery.Test
         [Test]
         public void TestAddition()
         {
-            Element X = Element.qElement - 1;
-            Element Y = X + X;
-            Element Z = Y - X;
+            FE X = FE.qElement - 1;
+            FE Y = X + X;
+            FE Z = Y - X;
             Assert.IsTrue(Z.Equals(X));
 
-            using IEnumerator<Element> set = Element.GetRandom().GetEnumerator();
+            using IEnumerator<FE> set = FE.GetRandom().GetEnumerator();
             for (int i = 0; i < 1000; i++)
             {
-                Element x = set.Current;
-                Element y = x + x + x + x;
-                Element z = y - x - x - x - x;
+                FE x = set.Current;
+                FE y = x + x + x + x;
+                FE z = y - x - x - x - x;
                 Assert.IsTrue(z.IsZero);
                 set.MoveNext();
             }
         }
 
         [Test]
-        public void TestInverse()
+        public void TestHotPath()
         {
-            using IEnumerator<Element> set = Element.GetRandom().GetEnumerator();
+            using IEnumerator<FE> set = FE.GetRandom().GetEnumerator();
             for (int i = 0; i < 1000; i++)
             {
-                Element x = set.Current;
+                FE current = set.Current;
+                FE x = set.Current * set.Current;
+                FE.Inverse(in current, out FE y);
+
+                Assert.IsTrue(current.Equals(x * y));
+            }
+        }
+
+        [Test]
+        public void TestInverse()
+        {
+            using IEnumerator<FE> set = FE.GetRandom().GetEnumerator();
+            for (int i = 0; i < 1000; i++)
+            {
+                FE x = set.Current;
                 if (x.IsZero)
                 {
                     set.MoveNext();
                     continue;
                 }
-                Element.Inverse(x, out Element y);
-                Element.Inverse(y, out Element z);
+                FE.Inverse(x, out FE y);
+                FE.Inverse(y, out FE z);
                 Assert.IsTrue(z.Equals(x));
                 set.MoveNext();
             }
@@ -64,18 +78,35 @@ namespace Nethermind.Field.Montgomery.Test
         [Test]
         public void TestInverseMultiplication()
         {
-            using IEnumerator<Element> set = Element.GetRandom().GetEnumerator();
+            using IEnumerator<FE> set = FE.GetRandom().GetEnumerator();
             for (int i = 0; i < 1000; i++)
             {
-                Element x = set.Current;
+                FE x = set.Current;
                 if (x.IsZero)
                 {
                     set.MoveNext();
                     continue;
                 }
-                Element.Inverse(x, out Element y);
-                Element.MultiplyMod(x, y, out Element z);
+                FE.Inverse(x, out FE y);
+                FE.MultiplyMod(x, y, out FE z);
                 Assert.IsTrue(z.IsOne);
+                set.MoveNext();
+            }
+        }
+
+        [Test]
+        public void ProfileMultiplication()
+        {
+            using IEnumerator<FE> set = FE.GetRandom().GetEnumerator();
+            for (int i = 0; i < 100000; i++)
+            {
+                FE x = set.Current;
+                if (x.IsZero)
+                {
+                    set.MoveNext();
+                    continue;
+                }
+                FE.MultiplyMod(x, x, out FE z);
                 set.MoveNext();
             }
         }
@@ -83,12 +114,12 @@ namespace Nethermind.Field.Montgomery.Test
         [Test]
         public void TestSerialize()
         {
-            using IEnumerator<Element> set = Element.GetRandom().GetEnumerator();
+            using IEnumerator<FE> set = FE.GetRandom().GetEnumerator();
             for (int i = 0; i < 1000; i++)
             {
-                Element x = set.Current;
+                FE x = set.Current;
                 Span<byte> bytes = x.ToBytes();
-                Element elem = Element.FromBytes(bytes.ToArray());
+                FE elem = FE.FromBytes(bytes.ToArray());
                 Assert.IsTrue(x.Equals(elem));
                 set.MoveNext();
             }
@@ -97,12 +128,12 @@ namespace Nethermind.Field.Montgomery.Test
         [Test]
         public void TestSerializeBigEndian()
         {
-            using IEnumerator<Element> set = Element.GetRandom().GetEnumerator();
+            using IEnumerator<FE> set = FE.GetRandom().GetEnumerator();
             for (int i = 0; i < 1000; i++)
             {
-                Element x = set.Current;
+                FE x = set.Current;
                 Span<byte> bytes = x.ToBytesBigEndian();
-                Element elem = Element.FromBytes(bytes.ToArray(), true);
+                FE elem = FE.FromBytes(bytes.ToArray(), true);
                 Assert.IsTrue(x.Equals(elem));
                 set.MoveNext();
             }
@@ -111,17 +142,17 @@ namespace Nethermind.Field.Montgomery.Test
         [Test]
         public void TestSqrt()
         {
-            using IEnumerator<Element> set = Element.GetRandom().GetEnumerator();
+            using IEnumerator<FE> set = FE.GetRandom().GetEnumerator();
             for (int i = 0; i < 1000; i++)
             {
-                Element x = set.Current;
-                if (Element.Legendre(x) != 1)
+                FE x = set.Current;
+                if (FE.Legendre(x) != 1)
                 {
                     set.MoveNext();
                     continue;
                 }
-                Element.Sqrt(x, out Element sqrtElem);
-                Element.Exp(sqrtElem, 2, out Element res);
+                FE.Sqrt(x, out FE sqrtElem);
+                FE.Exp(sqrtElem, 2, out FE res);
                 Assert.IsTrue(x.Equals(res));
                 set.MoveNext();
             }
@@ -130,13 +161,13 @@ namespace Nethermind.Field.Montgomery.Test
         [Test]
         public void TestMultiInv()
         {
-            Element[] values =
+            FE[] values = new[]
             {
-                Element.SetElement(1), Element.SetElement(2), Element.SetElement(3)
+                FE.SetElement(1), FE.SetElement(2), FE.SetElement(3)
             };
 
-            Element[] gotInverse = Element.MultiInverse(values);
-            Element?[] expectedInverse = NaiveMultiInverse(values);
+            FE[] gotInverse = FE.MultiInverse(values);
+            FE?[] expectedInverse = NaiveMultiInverse(values);
 
             Assert.IsTrue(gotInverse.Length == expectedInverse.Length);
             for (int i = 0; i < gotInverse.Length; i++)
@@ -145,12 +176,12 @@ namespace Nethermind.Field.Montgomery.Test
             }
         }
 
-        private static Element?[] NaiveMultiInverse(IReadOnlyList<Element> values)
+        static private FE?[] NaiveMultiInverse(IReadOnlyList<FE> values)
         {
-            Element?[] res = new Element?[values.Count];
+            FE?[] res = new FE?[values.Count];
             for (int i = 0; i < values.Count; i++)
             {
-                Element.Inverse(values[i], out Element x);
+                FE.Inverse(values[i], out FE x);
                 res[i] = x;
             }
             return res;